more on cli

+++ db/migrate/20200728220628_add_price_column_to_vehicles.rb
@@ -1,5 +1,5 @@
 class AddPriceColumnToVehicles < ActiveRecord::Migration[5.2]
   def change
-    add_column :vehicles, :price, :INTEGER
+    add_column :vehicles, :price, :float
   end
 end
